refactor: reorganiza diretorio de flags
