+++ src/features/common/shared/auth/index.ts
@@ -1,0 +1,2 @@
+export { propertyFromCookie } from "./modules/propertyFromCookie";
+export { propertyFromHeader } from "./modules/propertyFromHeader";
