+++ src/features/common/shared/auth/modules/propertyFromCookie/__tests__/propertyFromCookie.test.ts
@@ -1,6 +1,6 @@
 import { decode } from "next-auth/jwt";
 import type { ReadonlyRequestCookies } from "next/dist/server/web/spec-extension/adapters/request-cookies";
-import { propertyFromCookie } from "../propertyFromCookie";
+import { propertyFromCookie } from "..";
 
 jest.mock("next-auth/jwt", () => ({
     decode: jest.fn(),
