+++ src/features/common/shared/flags/index.ts
@@ -1,3 +1,3 @@
 export { diabetesNewProgram } from "./modules/diabetesNewProgram";
 export { hypertensionNewProgram } from "./modules/hypertensionNewProgram";
-export { developmentFlag } from "./modules/developmentFlag";
+export { print as developmentFlag } from "./modules/print";
