+++ src/features/common/shared/flags/modules/common/allowByMunicipalityId/index.ts
@@ -1,11 +1,9 @@
 import type { ReadonlyRequestCookies } from "next/dist/server/web/spec-extension/adapters/request-cookies";
-// import {
-//     municipalityIdSusFromCookie,
-//     municipalityIdSusFromHeader,
-// } from "./logic";
 import type { MunicipalityIdSus } from "./model";
-import { propertyFromHeader } from "../extractProperty/propertyFromHeader";
-import { propertyFromCookie } from "../extractProperty/propertyFromCookie";
+import {
+    propertyFromHeader,
+    propertyFromCookie,
+} from "@/features/common/shared/auth";
 export { buildDecide } from "../logic";
 export type { MunicipalityIdSus } from "./model";
